feat: CategoryDetailViewModel 분리진행
