+++ ios-unsplash/MainViewController/SearchView/CategoryDetailView/CategoryDetailView.swift
@@ -43,36 +43,6 @@ struct CategoryDetailView: View {
     }
 }
 
-final class CategoryDetailViewModel: ObservableObject {
-    @Published var photos: [Photo] = []
-    @Published var isLoading = false
-    
-    private var cancellables: Set<AnyCancellable> = []
-    private let repository: UnsplashRepository
-    
-    init(repository: UnsplashRepository) {
-        self.repository = repository
-    }
-    
-    func loadPhotos(query: String) {
-        isLoading = true
-        repository.searchPhotos(query: query, page: 1)
-            .receive(on: DispatchQueue.main)
-            .sink(receiveCompletion: { [weak self] completion in
-                self?.isLoading = false
-                switch completion {
-                case .failure(let error):
-                    print("\(error.localizedDescription)")
-                case .finished:
-                    print("photos")
-                }
-            }, receiveValue: { [weak self] response in
-                self?.photos = response.results
-            })
-            .store(in: &cancellables)
-    }
-}
-
 struct CategoryDetailView_Previews: PreviewProvider {
     static var previews: some View {
         //        CategoryDetailView(category: CategoryItem(name: "자연"), repository: UnsplashRepositoryImplementation(sessionProvider: URLSessionProviderImplementation()))
